Harmonise code and structure with other apps, use kpm-api-common
